rename selectors module to aoi_selectors to avoid collision with submodule.selectors
